backend/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
black==21.9b0
boto3==1.20.15
botocore==1.23.15
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.7
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
distlib==0.3.3
Django==3.2.8
django-cleanup==5.2.0
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-debug-toolbar-force==0.1.8
django-environ==0.8.1
django-filter==21.1
django-nine==0.2.5
django-storages==1.12.3
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
filelock==3.3.1
flake8==4.0.1
gunicorn==20.1.0
identify==2.3.1
idna==3.3
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.0.2
jmespath==0.10.0
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
oauthlib==3.1.1
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
pre-commit==2.15.0
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pycparser==2.20
pyflakes==2.4.0
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
regex==2021.10.23
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
Unidecode==1.3.2
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.9.0
zipp==3.6.0
